+++ Classes/ViewHelpers/Format/FileSizeViewHelper.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace T3v\T3vContent\ViewHelpers\Format;
 
 use T3v\T3vCore\ViewHelpers\AbstractViewHelper;
@@ -8,28 +9,39 @@ use T3v\T3vCore\ViewHelpers\AbstractViewHelper;
  *
  * @package T3v\T3vContent\ViewHelpers\Format
  */
-class FileSizeViewHelper extends AbstractViewHelper {
+class FileSizeViewHelper extends AbstractViewHelper
+{
   const SUFFIXES = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
+
+  public function initializeArguments()
+  {
+    $this->registerArgument('fileSize', 'int', 'The file size in bytes', true);
+    $this->registerArgument('base', 'int', 'The optional base unit to calculate the file size, defaults to `1000`', false, 1000);
+    $this->registerArgument('precision', 'int', ' The optional precision, defaults to `2`', false, 2);
+    $this->registerArgument('decimalSeparator', 'int', 'The optional decimal separator, defaults to `null`', false, null);
+    $this->registerArgument('thousandsSeparator', 'int', 'The optional thousands separator, defaults to `null`', false, null);
+
+  }
 
   /**
    * The view helper render function.
    *
-   * @param int $fileSize The file size in bytes
-   * @param int $base The optional base unit to calculate the file size, defaults to `1000`
-   * @param int $precision The optional precision, defaults to `2`
-   * @param int $decimalSeparator The optional decimal separator, defaults to `null`
-   * @param int $thousandsSeparator The optional thousands separator, defaults to `null`
    * @return string The formatted / human readable file size
    */
-  public function render($fileSize, $base = 1000, $precision = 2, $decimalSeparator = null, $thousandsSeparator = null) {
-    $fileSize           = intval($fileSize);
-    $base               = intval($base);
-    $precision          = intval($precision);
-    $decimalSeparator   = $decimalSeparator ?: $this->getDecimalSeparator();
+  public function render()
+  {
+
+    $fileSize = (int)$this->arguments['fileSize'];
+    $base = (int)$this->arguments['base'];
+    $precision = (int)$this->arguments['precision'];
+    $decimalSeparator = (int)$this->arguments['decimalSeparator'];
+    $thousandsSeparator = (int)$this->arguments['thousandsSeparator'];
+
+    $decimalSeparator = $decimalSeparator ?: $this->getDecimalSeparator();
     $thousandsSeparator = $thousandsSeparator ?: $this->getThousandsSeparator();
 
-    $pow      = floor(($fileSize ? log($fileSize) : 0) / log($base));
-    $pow      = min($pow, count(self::SUFFIXES) - 1);
+    $pow = floor(($fileSize ? log($fileSize) : 0) / log($base));
+    $pow = min($pow, count(self::SUFFIXES) - 1);
     $fileSize /= (1 << (10 * $pow));
 
     return sprintf(
@@ -44,7 +56,8 @@ class FileSizeViewHelper extends AbstractViewHelper {
    *
    * @return string The decimal separator
    */
-  protected function getDecimalSeparator() {
+  protected function getDecimalSeparator()
+  {
     $decimalSeparator = '.';
 
     switch ($this->getLanguage()) {
@@ -64,7 +77,8 @@ class FileSizeViewHelper extends AbstractViewHelper {
    *
    * @return string The thousands separator
    */
-  protected function getThousandsSeparator() {
+  protected function getThousandsSeparator()
+  {
     $thousandsSeparator = ',';
 
     switch ($this->getLanguage()) {