+++ Classes/ViewHelpers/SubheadlineLinkViewHelper.php
@@ -19,25 +19,31 @@ class SubheadlineLinkViewHelper extends AbstractTagBasedViewHelper {
    */
   protected $tagName = 'h2';
 
+  public function initializeArguments()
+  {
+    parent::initializeArguments();
+    $this->registerArgument('content', 'string', 'The content of the subheadline', false, null);
+    $this->registerArgument('layout', 'int', 'The optional layout (UID) of the headline, defaults to 100', false, 100);
+    $this->registerArgument('cssClass', 'string', 'The optional CSS class of the subheadline, defaults to `content-element__subtitle`', false, 'content-element__subtitle');
+  }
+
   /**
    * The view helper render function.
    *
-   * @param string $content The content of the subheadline
-   * @param int $layout The optional layout (UID) of the headline, defaults to `100`
-   * @param string $cssClass The optional CSS class of the subheadline, defaults to `ce__subtitle`
    * @return string The rendered subheadline
    */
-  public function render($content = null, $layout = 100, $cssClass = 'ce__subtitle') {
-    $content  = (string) $content;
-    $layout   = intval($layout);
-    $cssClass = (string) $cssClass;
+  public function render() {
+
+    $content = (string)$this->arguments['content'];
+    $layout = (int)$this->arguments['layout'];
+    $cssClass = (string)$this->arguments['cssClass'];
 
     if (empty($content)) {
       $content = $this->renderChildren();
     }
 
-    if (!empty($content) && $layout != 0 && $layout != 100) {
-      $layout  = $layout + 1;
+    if (!empty($content) && $layout !== 0 && $layout !== 100) {
+      $layout++;
       $tagName = "h{$layout}";
 
       $this->tag->setTagName($tagName);
@@ -51,24 +57,27 @@ class SubheadlineLinkViewHelper extends AbstractTagBasedViewHelper {
     }
   }
 
+
+
   /**
    * Creates a link.
    *
    * @param string $content The content of the link
    * @param string $href The optional href attribute of the link, defaults to `#`
    * @param string $title The optional title attribute of the link
-   * @return \TYPO3\CMS\Fluid\Core\ViewHelper\TagBuilder The link
+   * @return \TYPO3Fluid\Fluid\Core\ViewHelper\TagBuilder The link
    */
-  protected function createLink($content, $href = '#', $title = null) {
-    $content = (string) $content;
-    $href    = (string) $href;
-    $title   = (string) $title;
+  protected function createLink($content, $href = '#', $title = null)
+  {
+    $content = (string)$content;
+    $href = (string)$href;
+    $title = (string)$title;
 
     if (empty($title)) {
       $title = $content;
     }
 
-    $link = GeneralUtility::makeInstance('TYPO3\CMS\Fluid\Core\ViewHelper\TagBuilder');
+    $link = GeneralUtility::makeInstance(\TYPO3Fluid\Fluid\Core\ViewHelper\TagBuilder::class);
     $link->setTagName('a');
     $link->addAttribute('href', $href);
     $link->addAttribute('title', $title);