+++ Classes/ViewHelpers/SubheadlineViewHelper.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace T3v\T3vContent\ViewHelpers;
 
 use T3v\T3vCore\ViewHelpers\AbstractTagBasedViewHelper;
@@ -8,7 +9,8 @@ use T3v\T3vCore\ViewHelpers\AbstractTagBasedViewHelper;
  *
  * @package T3v\T3vContent\ViewHelpers
  */
-class SubheadlineViewHelper extends AbstractTagBasedViewHelper {
+class SubheadlineViewHelper extends AbstractTagBasedViewHelper
+{
   /**
    * The tag name.
    *
@@ -16,25 +18,32 @@ class SubheadlineViewHelper extends AbstractTagBasedViewHelper {
    */
   protected $tagName = 'h2';
 
+  public function initializeArguments()
+  {
+    parent::initializeArguments();
+    $this->registerArgument('content', 'string', 'The content of the subheadline', false, null);
+    $this->registerArgument('layout', 'int', 'The optional layout (UID) of the headline, defaults to 100', false, 100);
+    $this->registerArgument('cssClass', 'string', 'The optional CSS class of the subheadline, defaults to `content-element__subtitle`', false, 'content-element__subtitle');
+  }
+
   /**
    * The view helper render function.
    *
-   * @param string $content The content of the subheadline
-   * @param int $layout The optional layout (UID) of the headline, defaults to `100`
-   * @param string $cssClass The optional CSS class of the subheadline, defaults to `ce__subtitle`
    * @return string The rendered subheadline
    */
-  public function render($content = null, $layout = 100, $cssClass = 'ce__subtitle') {
-    $content  = (string) $content;
-    $layout   = intval($layout);
-    $cssClass = (string) $cssClass;
+  public function render()
+  {
+
+    $content = (string)$this->arguments['content'];
+    $layout = (int)$this->arguments['layout'];
+    $cssClass = (string)$this->arguments['cssClass'];
 
     if (empty($content)) {
       $content = $this->renderChildren();
     }
 
-    if (!empty($content) && $layout != 0 && $layout != 100) {
-      $layout  = $layout + 1;
+    if (!empty($content) && $layout !== 0 && $layout !== 100) {
+      $layout++;
       $tagName = "h{$layout}";
 
       $this->tag->setTagName($tagName);