+++ Classes/ViewHelpers/ViewColumnViewHelper.php
@@ -2,92 +2,43 @@
 namespace T3v\T3vContent\ViewHelpers;
 
 use T3v\T3vCore\ViewHelpers\AbstractViewHelper;
-
-use T3v\T3vContent\Service\ExtensionService;
+use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
 
 /**
  * The view column view helper class.
  *
  * @package T3v\T3vContent\ViewHelpers
  */
-class ViewColumnViewHelper extends AbstractViewHelper {
-  /**
-   * The extension service.
-   *
-   * @var \T3v\T3vContent\Service\ExtensionService
-   * @inject
-   */
-  protected $extensionService;
+class ViewColumnViewHelper extends AbstractViewHelper
+{
 
   /**
-   * The view helper render function.
-   *
-   * @param int $viewColumn The UID of the view column
-   * @param array $data The page data
-   * @return string The rendered content of the view column
+   * @return void
    */
-  public function render(int $viewColumn, $data) {
-    $output       = '';
-    $viewChildren = $data['tx_gridelements_view_children'];
-    $viewChildren = $this->filterViewChildrenByViewColumn($viewChildren, $viewColumn);
-
-    if ($this->extensionService->runningInStrictMode()) {
-      $viewChildren = $this->filterViewChildrenBySysLanguage($viewChildren);
-    }
-
-    foreach($viewChildren as $viewChild) {
-      $uid              = $viewChild['uid'];
-      $viewChildId      = "tx_gridelements_view_child_{$uid}";
-      $viewChildContent = $data[$viewChildId];
-
-      if (!empty($viewChildContent)) {
-        $output .= $viewChildContent;
-      }
-    }
-
-    return $output;
+  public function initializeArguments()
+  {
+    $this->registerArgument('viewColumn', 'int', 'The UID of the view column', true);
+    $this->registerArgument('data', 'array', 'The page data', true, null);
   }
 
   /**
-   * Filters the view children by a view column.
+   * The view helper render static function.
    *
-   * @param array $viewChildren The view children
-   * @param int $viewColumn The UID of the view column
-   * @return array The filtered view children
+   * @param array $arguments
+   * @param \Closure $renderChildrenClosure
+   * @param RenderingContextInterface $renderingContext
+   * @return mixed
    */
-  protected function filterViewChildrenByViewColumn(array $viewChildren, int $viewColumn) {
-    $result = [];
+  public static function renderStatic(array $arguments, \Closure $renderChildrenClosure, RenderingContextInterface $renderingContext)
+  {
+    $data = $arguments['data'];
+    $viewColumn = $arguments['viewColumn'];
+    $key = 'tx_gridelements_view_column_' . $viewColumn;
 
-    if ($viewChildren) {
-      foreach($viewChildren as $viewChild) {
-        if (intval($viewChild['tx_gridelements_columns']) == $viewColumn) {
-          $result[] = $viewChild;
-        }
-      }
+    if (array_key_exists($key, $data)) {
+      return $data[$key];
     }
 
-    return $result;
-  }
-
-  /**
-   * Filters the view children by the current system language.
-   *
-   * @param array $viewChildren The view children
-   * @return array The filtered view children
-   */
-  protected function filterViewChildrenBySysLanguage($viewChildren) {
-    $result = [];
-
-    if ($viewChildren) {
-      $sysLanguageUid = $this->getSysLanguageUid();
-
-      foreach($viewChildren as $viewChild) {
-        if (intval($viewChild['sys_language_uid']) == $sysLanguageUid) {
-          $result[] = $viewChild;
-        }
-      }
-    }
-
-    return $result;
+    return '';
   }
 }