+++ ext_emconf.php
@@ -1,37 +1,40 @@
 <?php
+/** @var string $_EXTKEY */
+
 $EM_CONF[$_EXTKEY] = [
-  'title' => 'T3v Content',
-  'description' => 'The content extension of TYPO3Voila.',
-  'author' => 'Maik Kempe',
-  'author_email' => 'mkempe@bitaculous.com',
-  'author_company' => 'Bitaculous - It\'s all about the bits, baby!',
-  'category' => 'fe',
-  'state' => 'stable',
-  'version' => '8.0.0',
-  'createDirs' => '',
-  'uploadfolder' => false,
-  'clearCacheOnLoad' => false,
-  'constraints' => [
-    'depends' => [
-      'typo3' => '9.5.0-10.4.99',
-      'fluid_styled_content' => '9.5.0-10.4.99',
-      'gridelements' => '',
-      'vhs' => '',
-      't3v_core' => ''
+    'title' => 'T3v Content',
+    'description' => 'The content extension of TYPO3Voila.',
+    'author' => 'Maik Kempe',
+    'author_email' => 'mkempe@bitaculous.com',
+    'author_company' => 'Bitaculous - It\'s all about the bits, baby!',
+    'category' => 'fe',
+    'state' => 'alpha',
+    'version' => '8.0.0',
+    'createDirs' => '',
+    'uploadfolder' => false,
+    'clearCacheOnLoad' => false,
+    'constraints' => [
+        'depends' => [
+            'typo3' => '9.5.0-10.4.99',
+            'fluid_styled_content' => '9.5.0-10.4.99',
+            'gridelements' => '',
+            'sms_responsive_images' => '',
+            'vhs' => '',
+            't3v_core' => ''
+        ],
+        'conflicts' => [
+            'css_styled_content' => '9.5.0-10.4.99'
+        ],
+        'suggests' => []
     ],
-    'conflicts' => [
-      'css_styled_content' => '9.5.0-10.4.99'
+    'autoload' => [
+        'psr-4' => [
+            'T3v\\T3vContent\\' => 'Classes'
+        ]
     ],
-    'suggests' => []
-  ],
-  'autoload' => [
-    'psr-4' => [
-      'T3v\\T3vContent\\' => 'Classes'
+    'autoload-dev' => [
+        'psr-4' => [
+            'T3v\\T3vContent\\Tests\\' => 'Tests'
+        ]
     ]
-  ],
-  'autoload-dev' => [
-    'psr-4' => [
-      'T3v\\T3vContent\\Tests\\' => 'Tests'
-    ]
-  ]
 ];