+++ ext_localconf.php
@@ -3,67 +3,47 @@ defined('TYPO3_MODE') or die();
 
 // === Variables ===
 
-$namespace           = 'T3v';
-$extensionKey        = $_EXTKEY;
-$extensionSignature  = \T3v\T3vCore\Utility\ExtensionUtility::signature($namespace, $extensionKey);
-$configurationFolder = \T3v\T3vCore\Utility\ExtensionUtility::configurationFolder($extensionKey);
-$iconsFolder         = \T3v\T3vCore\Utility\ExtensionUtility::iconsFolder($extensionKey);
+/** @var string $_EXTKEY */
 
-// === Page TSconfig ===
+$namespace = 'T3v';
+$extensionKey = $_EXTKEY;
+$extensionSignature = \T3v\T3vCore\Utility\ExtensionUtility::signature($namespace, $extensionKey);
+$tsConfigFolder = \T3v\T3vCore\Utility\ExtensionUtility::tsConfigFolder($extensionKey);
+$iconsFolder = \T3v\T3vCore\Utility\ExtensionUtility::iconsFolder($extensionKey);
 
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig("<INCLUDE_TYPOSCRIPT: source=\"{$configurationFolder}/TSconfig.tsconfig\">");
+// === TSconfig ===
 
-// === Content Elements ===
-
-// --- Spacer Content Element ---
-
-\TYPO3\CMS\Extbase\Utility\ExtensionUtility::configurePlugin(
-  $extensionSignature,
-
-  // The unique name of the content element in upper camel case
-  'Spacer',
-
-  // The array holding the controller-action-combinations that are accessible
-  [
-    // The first controller and its first action will be the default
-    'ContentElements\SpacerContentElement' => 'index'
-  ],
-
-  // The array of non-cachable controller-action-combinations (they must already be enabled)
-  [
-    // 'ContentElements\SpacerContentElement' => 'index'
-  ]
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPageTSConfig(
+    "<INCLUDE_TYPOSCRIPT: source=\"{$tsConfigFolder}/Page.tsconfig\">"
 );
 
 // === Backend ===
 
 if (TYPO3_MODE === 'BE') {
-  // --- Icons ---
+    // --- Icons ---
 
-  $iconRegistry = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Imaging\IconRegistry::class);
+    $iconRegistry = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Imaging\IconRegistry::class);
 
-  $icons = [
-    'Spacer Content Element'        => "{$iconsFolder}/ContentElements/SpacerContentElement.svg",
-    'Accordion Grid Element'        => "{$iconsFolder}/GridElements/AccordionGridElement.svg",
-    'Accord Grid Element'           => "{$iconsFolder}/GridElements/AccordGridElement.svg",
-    '2 Columns Layout Grid Element' => "{$iconsFolder}/GridElements/ColumnLayoutGridElement/2Columns.svg",
-    '3 Columns Layout Grid Element' => "{$iconsFolder}/GridElements/ColumnLayoutGridElement/3Columns.svg",
-    'Conditionally Grid Element'    => "{$iconsFolder}/GridElements/ConditionallyGridElement.svg",
-    'Document Section Grid Element' => "{$iconsFolder}/GridElements/DocumentSectionGridElement.svg",
-    'Panel Grid Element'            => "{$iconsFolder}/GridElements/PanelGridElement.svg",
-    'Timeline Grid Element'         => "{$iconsFolder}/GridElements/TimelineGridElement.svg",
-    'Event Grid Element'            => "{$iconsFolder}/GridElements/EventGridElement.svg",
-    'Viewport Grid Element'         => "{$iconsFolder}/GridElements/ViewportGridElement.svg"
-  ];
+    $icons = [
+        'Accordion Grid Element' => "{$iconsFolder}/GridElements/AccordionGridElement.svg",
+        'Accord Grid Element' => "{$iconsFolder}/GridElements/AccordGridElement.svg",
+        'Card Grid Element' => "{$iconsFolder}/GridElements/CardGridElement.svg",
+        'Slider Grid Element' => "{$iconsFolder}/GridElements/SliderGridElement.svg",
+        'Carousel Grid Element' => "{$iconsFolder}/GridElements/CarouselGridElement.svg",
+        'Timeline Grid Element' => "{$iconsFolder}/GridElements/TimelineGridElement.svg",
+        'Session Grid Element' => "{$iconsFolder}/GridElements/SessionGridElement.svg",
+        'Time Slot Grid Element' => "{$iconsFolder}/GridElements/TimeSlotGridElement.svg",
+        'Event Grid Element' => "{$iconsFolder}/GridElements/EventGridElement.svg"
+    ];
 
-  foreach ($icons as $name => $source) {
-    $iconIdentifier = \T3v\T3vCore\Utility\IconUtility::identifier($name);
-    $iconSignature  = \T3v\T3vCore\Utility\IconUtility::signature($extensionKey, $iconIdentifier);
+    foreach ($icons as $name => $source) {
+        $iconIdentifier = \T3v\T3vCore\Utility\IconUtility::identifier($name);
+        $iconSignature = \T3v\T3vCore\Utility\IconUtility::signature($extensionKey, $iconIdentifier);
 
-    $iconRegistry->registerIcon(
-      $iconSignature,
-      \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-      ['source' => $source]
-    );
-  }
+        $iconRegistry->registerIcon(
+            $iconSignature,
+            \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+            ['source' => $source]
+        );
+    }
 }